feat(http): implemented http/json service
The server constantly fetches and parses the data from it's data
source and returns a JSON formatted string upon request.

Currently the server returns JSON like
{
  bus: [ /* array of busses */ ],
  stop: [ /* array of stops */ ],
  timestamp: /* timestamp of last udpate */,
}

Where a bus is a object like
{
  id:
  title:
  address:
  longitude:
  latitude:
  iconColor:
  direction:
  speed:
  timestamp:
  lastStop:
  route:
}

and stop is an object like
{
  title:
  longitude:
  latitude:
  icon:
  route:
}

For the current data source, stops seems to be always empty. Also certain
fields are never filled with actual values.

The current update interval for the server is set to 15000, because the
data source's processing time for a single request seems rather long (>1 sec).